:truck: [#1035] move the check_broken_links logic to a seperate file and also moved the check_broken_links tests to it's own file.
